Bug 853: Use consistent standards/rates for examples and tests that use Wi-Fi
The change to consider mandatory rates for control responses if a
suitable basic rate could not be found broke the tests involving the
wireless/wifi-simple-adhoc-grid example, and the
routing-olsr-regression testsuite.

In these cases it is actually the tests that need updating; in the
wifi-simple-adhoc-grid test and the bug780-test part of the
routing-olsr-regression testsuite, 802.11b rates were being used with
the Wi-Fi PHY standard left at the default of 802.11a. The fix here is
to set the standard to 802.11b in the test.

In the tx-regression-test part of routing-olsr-regression, the problem
was that the reference traces needed to be updated given the control
response rate fix.  In this case I have also modified the test to set
the standard to 802.11a for clarity.