dca-txop: implement queue switching for TX radio:
1) Add two attributes: ChannelActiveTimeout and
QueueUtilizationInterval. The first one decides how long to keep
sending on a channel before disabling it. Default is 2 seconds.
The second one is to decide how long to keep transmitting on a
switched to channel before switching again. Default is 250 ms.

2) Implement a linked list with current "active" channels.
Queue switcher switches between them when transmitting.

3) Start the queue switching ordeals here when transmitting
the first packet on the TX radio